Trait pallet_liquidity_rewards::CurrencyGroupChange

pub trait CurrencyGroupChange {
    type GroupId;
    type CurrencyId;

    fn attach_currency(
        currency_id: Self::CurrencyId,
        group_id: Self::GroupId
    ) -> Result<(), DispatchError>;
    fn currency_group(currency_id: Self::CurrencyId) -> Option<Self::GroupId>;
}

Support for change currencies among groups.

Required Associated Types

type GroupId

Type used to identify the group

type CurrencyId

Type used to identify the currency

Required Methods

fn attach_currency(
currency_id: Self::CurrencyId,
group_id: Self::GroupId
) -> Result<(), DispatchError>

Associate the currency to a group. If the currency was previously associated to another group, the associated stake is moved to the new group.

fn currency_group(currency_id: Self::CurrencyId) -> Option<Self::GroupId>

Returns the associated group of a currency.