Crate pallet_liquidity_rewards

Rewards Pallet

The Rewards pallet provides functionality for distributing rewards to different accounts with different currencies. The distribution happens when an epoch (a constant time interval) finalizes. The user can stake an amount during one of more epochs to claim the reward.

Rewards pallet can be configured with any implementation of cfg_traits::rewards traits which gives the reward behavior.

The Rewards pallet provides functions for:

Stake/Unstake a currency amount.

Claim the reward given to a staked currency.

Admin methods to configure epochs, currencies and reward groups.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

transactional

Provides functionality around the transaction storage.

weights

Autogenerated weights for pallet_liquidity_rewards

Structs

BoundedBTreeMap

A bounded map based on a B-Tree.

EpochChanges

Type that contains the stake properties of stake class

EpochData

Type that contains the associated data of an epoch

EpochTimestamp

Type that contains the timestamp of an epoch.

FirstEpochTimestamp

Traits

AccountRewards

Abstraction over a distribution reward system for accounts.

CurrencyGroupChange

Support for change currencies among groups.

DistributedRewards

Distribution mechanisms over group rewards. This trait is implemented automatically for all GroupRewards with the requested bounds.

EnsureAdd

Performs addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureAddAssign

Performs self addition that returns ArithmeticError instead of wrapping around on overflow.

GroupRewards

Abstraction over a distribution reward groups.

Type Definitions

DomainIdOf

Attribute Macros

transactional

Execute the annotated function in a new storage transaction.