Trait pallet_liquidity_rewards::GroupRewards

pub trait GroupRewards {
    type Balance;
    type GroupId;

    fn reward_group(
        group_id: Self::GroupId,
        reward: Self::Balance
    ) -> Result<(), DispatchError>;
    fn group_stake(group_id: Self::GroupId) -> Self::Balance;
}

Abstraction over a distribution reward groups.

Required Associated Types

type Balance

Type used as balance for all currencies and reward.

type GroupId

Type used to identify the group

Required Methods

fn reward_group(
group_id: Self::GroupId,
reward: Self::Balance
) -> Result<(), DispatchError>

Reward a group distributing the reward amount proportionally to all associated accounts. This method is called by distribution method only when the group has some stake.

fn group_stake(group_id: Self::GroupId) -> Self::Balance

Retrieve the total staked amount.